Move logic to CLI package
